Renamed: component PickerCurrentPage.vue
