+++ apps/frontend/src/components/page-products/PickerCurrentPage.vue
